+++ UpgradeWorld/Executor.cs
@@ -8,7 +8,6 @@ namespace UpgradeWorld;
 public static class Executor
 {
   private static readonly List<ExecutedOperation> operations = [];
-  private static readonly List<Action> cleanUps = [];
   private static Coroutine? executionCoroutine;
   private static MonoBehaviour? context;
   public static void SetUser(ZRpc? user)
@@ -30,53 +29,42 @@ public static class Executor
   public static void StopExecution()
   {
     if (context == null) throw new Exception("Executor context is not set. Call Executor.SetContext from a MonoBehaviour before stopping execution.");
+    operations.Clear();
+    // Needed to indicate end of generation for some mods.
+    if (Hud.instance)
+      Hud.instance.m_loadingIndicator.SetShowProgress(false);
+
     if (executionCoroutine == null) return;
     context.StopCoroutine(executionCoroutine);
     executionCoroutine = null;
   }
   public static void AddOperation(ExecutedOperation operation, bool autoStart)
   {
-    operation.Init();
+    bool start = Settings.AutoStart || autoStart;
+    if (!operation.Init(start))
+      return;
     operations.Add(operation);
 
-    if (executionCoroutine == null && (Settings.AutoStart || autoStart))
+    if (executionCoroutine == null && start)
       StartExecution();
   }
-  public static void AddCleanUp(Action cleanUp)
+
+  public static List<ExecutedOperation> GetOperations()
   {
-    cleanUps.Add(cleanUp);
+    return operations;
   }
 
-  private static void DoClean()
-  {
-    foreach (var cleanUp in cleanUps) cleanUp();
-    cleanUps.Clear();
-    StopExecution();
-  }
-  public static void RemoveOperations()
-  {
-    operations.Clear();
-    DoClean();
-    // Needed to indicate end of generation for some mods.
-    if (Hud.instance)
-      Hud.instance.m_loadingIndicator.SetShowProgress(false);
-  }
   private static IEnumerator ExecuteCoroutine()
   {
     var sw = Stopwatch.StartNew();
-    while (true)
+    while (operations.Count > 0)
     {
-      if (operations.Count == 0)
-      {
-        sw.Stop();
-        DoClean();
-        yield break;
-      }
       sw.Restart();
       yield return operations[0].Execute(sw);
-
       operations.RemoveAt(0);
     }
+    sw.Stop();
+    StopExecution();
   }
   public const long ProgressMin = 100; // 0.1 seconds
   public const int ZdoMaxUpdates = 10000;