+++ UpgradeWorld/actions/base/ZoneOperation.cs
@@ -36,9 +36,6 @@ public abstract class ZoneOperation(Terminal context, FiltererParameters args) :
     {
       var zone = ZonesToUpgrade[ZoneIndex];
       var success = ExecuteZone(zone);
-      // Makes the zone unload as soon as possible.
-      if (success && ZoneSystem.instance.m_zones.TryGetValue(zone, out var zoneObj))
-        zoneObj.m_ttl = ZoneSystem.instance.m_zoneTTL;
       MoveToNextZone(success);
 
       if (success)