+++ UpgradeWorld/actions/objects/RemoveObjects.cs
@@ -7,13 +7,16 @@ public class RemoveObjects(Terminal context, IEnumerable<string> ids, DataParame
 
   protected override bool ProcessZDO(ZDO zdo)
   {
+    // Have to check if already removed.
+    if (zdo.m_prefab == 0)
+      return false;
     Helper.RemoveZDO(zdo);
     return true;
   }
 
   protected override string GetNoObjectsMessage() => "No objects found to remove.";
 
-  protected override string GetInitMessage() => $"Removing {TotalCount} object{(TotalCount > 1 ? "s" : "")}";
+  protected override string GetInitMessage() => $"Removing {TotalCount} object{(TotalCount > 1 ? "s" : "")}.";
 
   protected override string GetProcessedMessage() => $"Removed: {ProcessedCount}";